Added the Automated Trading Platform as described in https://bitcointalk.org/index.php?topic=109317.0
This is not an official part of OpenPay.  It was originally intended as a test case for the exchange connector and sort of grew from there.
DO NOT RUN THE TRADING PLATFORM UNLESS YOU KNOW WHAT YOU ARE DOING.
ALTHOUGH IT IS HIGHLY UNLIKELY TO HAPPEN, IN A WORST CASE SCENARIO IT COULD BANKRUPT YOUR MTGOX ACCOUNT WITH A FEW ERRANT TRADES

The author hereby disclaims any liability.

To start the application run "java -jar isis.jar"